refactor: move getUserSubHandle to lib/users
