+++ src/components/messenger/lib/utils.test.ts
@@ -1,4 +1,4 @@
-import { getUserSubHandle, highlightFilter } from './utils';
+import { highlightFilter } from './utils';
 
 describe('highlightFilter', () => {
   it('returns unmodified text when filter is an empty string', () => {
@@ -56,22 +56,5 @@ describe('highlightFilter', () => {
     expect(result[1].type).toEqual('span');
     expect(result[1].props.children).toEqual('lo');
     expect(result[2]).toEqual(' World');
-  });
-});
-
-describe('getUserSubHandle', () => {
-  it('returns primaryZID when it is present', () => {
-    const user = { primaryZID: 'zid123', primaryWallet: { id: 'wallet-id-1', publicAddress: 'address456' } };
-    expect(getUserSubHandle(user.primaryZID, user.primaryWallet.publicAddress)).toEqual('zid123');
-  });
-
-  it('returns truncated publicAddress from the first wallet when primaryZID is absent', () => {
-    const user = { primaryZID: null, primaryWallet: { id: 'wallet-id-1', publicAddress: '0x123456789' } };
-    expect(getUserSubHandle(user.primaryZID, user.primaryWallet.publicAddress)).toEqual('0x1234...6789');
-  });
-
-  it('returns empty string when both primaryZID and wallets are absent', () => {
-    const user = { primaryZID: null, primaryWallet: null };
-    expect(getUserSubHandle(user.primaryZID, user.primaryWallet)).toEqual('');
   });
 });
