+++ src/components/messenger/list/group-management/container.tsx
@@ -22,7 +22,7 @@ import { GroupManagementErrors, EditConversationState } from '../../../../store/
 import { User, denormalize as denormalizeChannel } from '../../../../store/channels';
 import { currentUserSelector } from '../../../../store/authentication/selectors';
 import { RemoveMemberDialogContainer } from '../../../group-management/remove-member-dialog/container';
-import { getUserSubHandle } from '../../lib/utils';
+import { getUserSubHandle } from '../../../../lib/user';
 
 export interface PublicProperties {
   searchUsers: (search: string) => Promise<any>;
