+++ src/lib/user.test.ts
@@ -1,4 +1,4 @@
-import { displayName } from './user';
+import { displayName, getUserSubHandle } from './user';
 
 import { User } from '../store/channels';
 
@@ -22,6 +22,23 @@ describe('User Lib', () => {
 
     it('returns "Unknown" if user is not found', () => {
       expect(displayName(null)).toEqual('Unknown');
+    });
+  });
+
+  describe(getUserSubHandle, () => {
+    it('returns primaryZID when it is present', () => {
+      const user = { primaryZID: 'zid123', primaryWallet: { id: 'wallet-id-1', publicAddress: 'address456' } };
+      expect(getUserSubHandle(user.primaryZID, user.primaryWallet.publicAddress)).toEqual('zid123');
+    });
+
+    it('returns truncated publicAddress from the first wallet when primaryZID is absent', () => {
+      const user = { primaryZID: null, primaryWallet: { id: 'wallet-id-1', publicAddress: '0x123456789' } };
+      expect(getUserSubHandle(user.primaryZID, user.primaryWallet.publicAddress)).toEqual('0x1234...6789');
+    });
+
+    it('returns empty string when both primaryZID and wallets are absent', () => {
+      const user = { primaryZID: null, primaryWallet: null };
+      expect(getUserSubHandle(user.primaryZID, user.primaryWallet)).toEqual('');
     });
   });
 });
